pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ott-monitor"
version = "0.1.0"

dependencies = [
    "ffmpeg-python",
    "python-dotenv",
]